Fixed warnings

+++ B-I-K-E-Shuffler/lib/ShufflerLibrary/ShufflerConstants.hpp
@@ -53,7 +53,7 @@ constexpr float BELT_MM_PER_REVOLUTION = 88.0f; // Find this value by testing, t
 constexpr int BELT_STEPS_PER_MM = BELT_STEPS_PER_REVOLUTION / BELT_MM_PER_REVOLUTION; // This number should be pretty large
 
 constexpr int BELT_LENGTH_MM = 1000; // Find this value by testing, the length of the belt in mm
-constexpr int BELT_LENGTH_STEPS = BELT_LENGTH_MM * BELT_STEPS_PER_MM;
+constexpr long BELT_LENGTH_STEPS = BELT_LENGTH_MM * BELT_STEPS_PER_MM;
 
 constexpr int BELT_DEFAULT_POSITION = 0; // TODO:: Find this value using calibration method
